Move function tests outside Netlify functions directory
