+++ package.json
@@ -13,7 +13,7 @@
     "build:preview": "npm run build:hugo:preview",
     "build:hugo": "hugo -d ../dist -s site",
     "build:hugo:preview": "npm run build:hugo -- -D -F",
-    "test:functions": "node --test netlify/functions/create-checkout-session.test.js netlify/functions/stripe-webhook.test.js",
+    "test:functions": "node --test tests/netlify-functions/create-checkout-session.test.js tests/netlify-functions/stripe-webhook.test.js",
     "postbuild": "node algolia-indexer.js"
   },
   "author": "",
